Two bugs fixed:
(1) First we re-initialize the sText control each time the semanticDataProvider is changed. This bug re-surfaced when we put a sText, within an itemRenderer part of a list that scrolls.

(2) Then each time that a bound-variable is found, we add it to the list of of targetAttributes. That way, we consider any bound-variable as target attributes defined by the user.

These two issues have been found when we used the sText component within an HorizontalList item renderer. Each time that the renderer was recycled by the Flex engine, the component was not updated. And once it was updated by fixing (1), then it was removing the component because the "text" attribute of the component was the value of the previous rendering of the component, and not the bound-variable anymore (because it as been replaced). It is what we fixed with (2).