Make sure that if "source" is not empty, but that there are "targetAttributes" that we continue to process the initialization of the sImage control.
The reason why the source of a sImage will only be used directly if no targetAttributes are defined is simple. If the sImage is part of an itemRenderer that get recycled when a list scroll, then the source that was empty is not empty anymore when it gets recycled. Since there were targetAttributes, the control won't remove itself. If this rule won't be in place, then it means that the images would be repeated one after the other, in a loop, when the list would get scrolled.