This is a major upgrade of the sWebMap component. It includes this series of new options and behaviors:
(1) [disableBrowseMode] A new option to disable the "browse" mode. This mode is used not to list all the results on an empty search string.

(2) [disableBrowseMode] A new option to display the filters by default (used when disableBrowseMode = true)

(3) [queryFocusWindowOnly] A new option to specify that only the focus windows will be queried )used when enableFocusWindow = true)

(4) [focusMapShowZoomControl] A new option to specify if we want to show the zooming control on the focus maps

(5) [focusMapShowPanControl] A new option to specify if we want to show the panning control on the focus maps

(6) [focusMapShowMapTypeControl] A new option to specify if we want to show the map type selection control on the focus maps

(7) [focusMapShowStreetViewControl] A new option to specify if we want to show the street view control on the focus maps

(8) [paginatorLocation] A new option to specify where we want the paginator control (on top or at the bottom of the results list)

(9) [minimalFocusWindowRectangleRadius] A new option to specify the minimum radius of the rectangles shown on the main map that represents the focus maps.

(10) [enableDirectionsService] A new option to enable the "get directions" feature between two markers displayed on the map

(11) [mapTypeIds] A new option to list all the map types we want to use for the sWebMap control (the custom, and built-in ones in the Google API).

(12) [directionsShowTrafficLayer] A new option to show a checkbox to enable/disable the traffic overlay layer when finding the path between two markers.

(13) [mapTypes] A new option to define new custom map types for the control

(14) [mapLayers] A new option to defome new custom overlay for the control

(15) [searchPlugins] A new option to define new pre-defined search spaces for the user (this option does create a combo box at the right side of the Search button